feat: 🎸 suggest location name, split some logics into hooks (#442)
